requirements.txt
# SPDX-FileCopyrightText: 2022 Diego Elio Pettenò
#
# SPDX-License-Identifier: 0BSD

click
cmarkgfm
jinja2
pygal_maps_world
